requirements.txt
DateTime==4.3
decorator==4.2.1
matplotlib==2.2.2
numpy==1.14.3
numpydoc==0.8.0
pandas==0.23.0
pandocfilters==1.4.2
path.py==11.0.1
pathlib2==2.3.2
pathlib>=1.0.1
pprint>=0.1
pylint
pycodestyle
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
xlrd==1.1.0
XlsxWriter==1.0.4
python-pptx>=0.6.17
fix_yahoo_finance==0.1.33
